Prepare for transition to new architecture (#42)
This PR prepares some more project infrastructure for the transition to
the new architecture. Most notably, it introduces Pylint on the example
of the new supervisor, such that we write good new code without having
to worry about pylint-compliance of the old code. Furthermore, we fail
on low coverage for the new modules.