Storage Component Overhaul (#55)
Major storage component overhaul. 

Implemented notion of filesystem and files, samples and a storage database to keep track of datasets, files and samples. Please refer to storage README on more details on the inner workings of this new storage component.

Resolves #11 and #44.

With contributions to #43, #47, #32 (in the storage component and as a
draft on how to implement the other components), #37, and #51.

Co-authored-by: Maximilian Böther <maxi+git@boether.de>