Introduce conda environment, fix Docker setup, test Docker in CI, cache in CI, docker-compose, small fixes (#67)
This changes our dependency structure to a conda environment. All
backend components now share a conda environment. Only the
dev-dependencies are held in a dev-requirements.txt file because its
easier to install in CI, if required. We use this conda environment in
CI (with caching, yay). Furthermore, I move the Docker structure outside
of the actual Python module. We introduce docker-compose integration
testing in CI and docker-compose for a local setup of Modyn. Last, this
PR fixes some small issues that prevented the storage service from being
run.

I decided to move an actual example integration test for the Storage to
the next PR.

Progress towards #54 and #43.
Closes #31.