Refactoring of triggering and Supervisor-Storage communication; Outline of Supervisor-Selector and Supervisor-Trainer communication (#79)
This PR changes how the triggering algorithms within the Supervisor
work. We now unify handling of experiment mode (replay) and the
production mode by internally batching new_data. We inform the Selector
about new data when there is no trigger. If there is a trigger, we
inform the Selector about the data up to (and including) the data point
that caused the trigger. Then, we obtain a trigger ID from the Selector,
confirming that the Selector has acknowledged the trigger. Then, we tell
the GPU node about the trigger ID and the pipeline ID. With that
information, the GPU node will be able to query all the data from the
Selector that it should train on.

Progress towards #10, #78, #64, #51.