Add TimeTrigger (#154)
This PR adds a TimeTrigger to the Supervisor. With the TimeTrigger, we can trigger every X seconds/minutes/hours/days.

For this, we add support for empty triggers, i.e., time passed but no
new data point in that time has arrived.

This trigger is relevant for our Criteo benchmark.