Fix docker-compose access denied error + Speedup of Docker build process (#227)
Finally, a fix for this very annoying problem :D

We do not rely on docker-compose anymore for building the base image but
instead build the base image ourselves. Furthermore, we change the
Dockerfile such that changes in Modyn's sources will not lead to
re-download of all dependencies, which will greatly speedup iteration
time. Only if the environment.yml is changed, then the process will be
retriggered.

The only downside is that `pip install -e .` needs to be run manually,
but I suppose we can live with that.