Introduction of Abstract Downsampling Strategy (#230)
This PR introduces the support for more advanced data selection policies
and implements one of them, the LossBasedDownsample.

In the Selector, there are several new classes. The first one is
AbstractDownsampleStrategy and is used to presample the dataset directly
on the database (instead of working on the full database, we just use a
percentage of it). Then two methods are left abstract for concrete
implementations. These two methods are used to inform the TrainerServer
of what it must do to select the samples.

Also in the TrainerServer several classes have been added. These classes
are called Remote* where * is the name of the corresponding class in the
selector. The class in the selector is used to build the corresponding
class in the trainer server, which is actually used to sample the
points. Sampling is done in the trainer server since it requires metrics
that are computed during training (gradient norm, loss..).

The "protocol" is the following:

during PytorchTrainer's init it asks to the selector if a remote
sampling strategy is used and, if yes, what is the name and which
parameters must be used to instantiate it
the trainer server instantiates the class passing the parameters
received from the selector and other local parameters (for example the
loss fct and the model)
Then, during the training loop, if remote downsampling is enabled the
function .sample(...) is called and the points are sampled accordingly

---------

Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>