Add Model Storage (#232)
Implementing model storage component based on #167. Currently, trained
models are stored on the supervisor which are downloaded from the
trainer server. We want to have a separate component storing the models.
For the supervisor it should then be possible to continue training from
a stored model or to evaluate it.

---------

Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>