Add CSVFileWrapper (#273)
In this PR, the CSV file wrapper is introduced.

Bytes are decoded to strings, the required rows are extracted and labels
are separated from samples. The labels are returned as integers, the
samples are reassembled in csv-like format and converted to bytes.
Pandas is not needed.

The bytes_parser_function takes care of decoding the bytes, splitting
the string to separators and converting/casting the data.