Introduce a super inefficient per class dataloader (#288)
To support some downsampling strategies, we need to be able to get
samples label by label. In this PR, an inefficient dataloader is
proposed.
These are the updates in this PR:

- new `PerClassOnlineDataset` in which samples are yielded only if
`label == self.filtered_label`. All the other samples are ignored. This
attribute can be changed to filter on a different label
- support in the sample then batch function (grpc call to get all the
distinct labels, _for_ loop to work label by label ..)
- support in the AbstractRemoteDownsampler (attribute
`requires_data_label_by_label` and abstract function
`_downsampler.inform_end_of_current_label` to handle the end of the
label