Refactor CoresetStrategy (#277)
To remove duplicate code and allow for more flexibility, the existing
downsampling strategies have been refactored.

Prior to this PR, the way to use Loss as downsampling strategy was
```
-name: LossDownsamplingStrategy
-config: .....
```

Now you can use
```
-name: CoresetStrategy
-config:
           - downsampling_config:
                   -  strategy: "Loss"
                   -  ratio: 50
                   - sample_then_batch: True
           - presampling_config: 
                   -  strategy: "Empty"
                   -  ratio: 40
           - .... 
```

Where downsampling_strategy inherits from `AbstractDownsamplingStrategy`
and presampling_strategy from `AbstractPresamplingStrategy`. In this
way, you can combine as you wish presampling and downsampling while
before presampling was always Random.

New presampling strategies can be introduced with this refactor such as
`RandomNoReplacement`, `RandomClassBalanced`, `RandomTriggerBalanced`
and so forth. A presampling strategy is just a SQL query that returns
the sampled ids. Also you can combine different presampling strategies
simply using the SQL union statement. For example, you can train on all
the points belonging to the last trigger plus a randomly presampled
version of the whole dataset.

Also, you can combine different downsampling strategy using the
`SchedulerDownsamplingStrategy`, where you use the `first_strategy`
during the first `threshold` triggers and then the `second_strategy`. In
the future, the switch might be performance based (if you are below a
certain accuracy you can use a strategy and if above another one).
For example, the scheduler can be useful in the following two scenari:
- start with all Data, then do GradNorm: in the original paper they
mention that, at the beginning of training, the gradient norm is almost
the same for all datapoints (all high loss since the weights are
randomly initiated) so it's like sampling from a uniform distribution
(no gain in computing the score, just use a stricter presampling)
- start with STB, then switch to BTS: sample then batch might be very
memory consuming (since you need to keep the score of all datapoints in
memory) when the dataset grows. Hence, we can stop using it after some
triggers and switch to the cheaper BTS

For example, the following pipeline produces the training pattern shown
below
```
downsampling_config:
  strategy: "Scheduler"
  threshold: 3
  first_strategy:
    strategy: "Empty"
  second_strategy:
    ratio: 50
    sample_then_batch: True
    strategy: "Loss"
```

![image](https://github.com/eth-easl/modyn/assets/59933812/f83e93d7-1c07-4f3e-b125-e4946ca3efe0)

A similar approach can be used in the `SchedulerPresamplingStrategy`