Wild Time Benchmarks (#285)
Introduce benchmarks [WildTIme](https://github.com/huaxiuyao/Wild-Time/tree/main). Some
helper methods are taken from [Wild-Time-data](https://github.com/wistuba/Wild-Time-Data)

In this PR, we introduce: - scripts for downloading data and formatting them so that they can be
used by Modyn - the models used for the 4 benchmarks (the same as those used in the
original paper) - the `tokenizer` option in the pipeline to support NLP tasks. Using `bytes_parser_function`, a new `tokenizer` is instantiated each time (with GET to huggingface for each sample) and the training is obviously
much slower (x15 or so based on a few runs). Using `tokenizer`, the `tokenizer` is instantiated only once and called sample by sample