Introduce policies from DeepCore (#304)
With this PR, we introduce several new downsampling methods: `Craig`,
`GradMatch`, `Submodular` (FacilityLocation, GraphCut and
LogDeterminant), `KcenterGreedy` and `Uncertainty` (Margin, Entropy, and
LeastConfidence). These methods are adapted from
[DEEPCORE](https://github.com/PatrickZH/DeepCore)

The following functionalities are implemented:
- `CoresetSupportingModule` support in both BTS and STB mode. Embeddings
are registered and provided in the `inform_samples` method if required.
- addition of the `device` in the constructor of the downsampling
methods. For example, kcenter uses a different implementation depending
on whether it runs on CPU or GPU.
- implementation of the techniques mentioned above. Many of which,
having a common behaviour, are abstracted with the
`AbstractMatrixDownsamplingStrategy` class
- fixing some bugs found by running the experiments (typically moving
weights to the correct device)

All methods were tested by comparing the results obtained with deepcore
in various controlled experiments. The available tests serve this
purpose.