Rewrite Storage Component in C++ (#270)
In the interest of #197 and #147 we have reworked the storage component
from Python to cpp. In the process of this we have dramatically
overworked the workflows involved while maintaining the logic and
functionality of the component.

For additional testing we extended the pipeline workflow with additional
c++ testing and linting. Note that timing changes from milliseconds to
seconds in the database.

Co-authored-by: Maximilian Böther <maxi+git@boether.de>
Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>