Fix broken args eval_dir on modyn-client (#374)
## What is the current case
Currently, the `eval_dir` arg on modyn-client is practically unusable:
it is a required arg; the client throws an exception if this path does
not exist on host system; supervisor throws an exception if this same
path does not exist on supervisor's container. Therefore user can
practically set nothing to `eval_dir` instead `.` (a reference to the
current working directory, possibly the only value that can pass both
checks)

Then users need to use `docker cp <more args>` to manually copy
evaluation files from supervisor container to host.

## What it would be after this PR

Supervisor has a default location, `eval_directory` to store evaluation
results (on supervisor container), whose value is in
`modyn_config.yaml`. It is up to the users to decide how to map this
path to host filesystem, e.g. via volume mapping or named volume in
`docker-compose.yaml`. The `eval_dir` arg on modyn-client is removed.