Fix Wildtime benchmark_utils create_timestamp() timezone to be UTC (#382)
# Bug
The `create_timestamp()` function in
`benchmark/wildtime_benchmarks/benchmark_utils.py` uses the **local**
timezone. The generated timestamp thus is **not UTC**.
However, modyn Storage assumes a UTC timestamp so that 1970/1/1
corresponds to 0.
Therefore, if the files are generated in Zurich local time, then
1970/1/1 is converted to -3600 in the `files` table.

# Fix
Fix tzinfo to be UTC timezone in `create_timestamp()`.