Remove old matrix eval implementation (#413)
This PR can be seen as a revert to
https://github.com/eth-easl/modyn/pull/334, which removes the current
matrix evaluation implementation.

This PR is part of https://github.com/eth-easl/modyn/pull/407. I singled
it out for easier review. After this PR, we can merge the new eval
implementation!

Although the number of involved files is large, all of them are just
removals so it's easy to understand the changes.