Make the reason for evaluation failure explicit (#425)
This is the last irrelevant PR I singled out from
https://github.com/eth-easl/modyn/pull/422.

One evaluation request can fail to start due to many reasons. Previously
the reasons are not explicitly reflected. This has two downsides:
1. In unit test when we want to test that "condition A" must be
satisfied to start evaluation, we can only assert the evaluation is not
started but the failure could be because of "condition B", i.e. we have
no way to verify the failure is because of "condition A".
2. On supervisor side, we don't know why the evaluation fails.

This PR makes the failure reason explicit.