Make `start_timestamp` and `end_timestamp` in evaluation gRPC requests optional (#428)
This PR makes two parameters I introduced in a previous PR
https://github.com/eth-easl/modyn/pull/388 optional. Previously, I made
it so that the default value "0" means no limit. I.e. a request with
both `start_timestamp` and `end_timestamp` as `0` means no time
restriction on the dataset.

I realized that this semantic is clumsy as sometimes we need "0" to mean
the real timestamp "0". And now there is conflict. This PR makes
`start_timestamp` and `end_timestamp` optional, and when they are
optional, it means there is no limit. When they are set with a value, it
means a limit with the numerical value.