Reject unknown properties in the pipeline config (#436)
Some configs in the pipeline config are optional, for example
`num_samples_to_pass`. If it is not specified, training is limited by
another config `epochs_per_trigger`.

It is common for users to make a typo. I once typed
`num_sample_to_pass`. The current validation allows additional
properties unspecified in the Pydantic model. This leads to the
consequence that the pipeline with typo `num_sample_to_pass` passed
through, and is executed as if `num_samples_to_pass` is not specified
and `epochs_per_trigger` is used. It wastes me one day to find it out.

In this PR I made it that unknown properties are rejected by default.
Then some outdated settings are discovered.

This PR also adds back the `dataset id` duplication check.