Change parameter list of the downsampling constructor and outlines the implementation of RHO-LOSS (#445)
The number of changed files is high, but allmost all of them are
function-signature-change, so it should be easy tor review!

This PR also outlines what will be done in the next several PRs by the
comments here
https://github.com/eth-easl/modyn/blob/2968ebd6c1178caf8c2c7e12456a1bb8de5564c9/modyn/selector/internal/selector_strategies/downsampling_strategies/rho_loss_downsampling_strategy.py#L11.

### Reason to place IL training logic inside the downsampler from
selector

This IL model training process is internal to the selection strategy
logic. It is not (a part of) a pipeline stage. The pipeline executor
should stay as much oblivious of training IL model as possible.