Accumulate batches in Batch-Then-Sample mode (#450)
Until now, in BtS mode, we did the forward pass on a batch, downsampled
it, then trained on it. This leads to different training dynamics (e.g.,
learning rate) because the batch size is smaller than in StB mode (or
training without sampling) and is also more inefficient since we do the
backward pass on smaller batches. With these changes, we accumulate the
forward pass data instead and do the backward pass as soon as we reach
the original batch size.