feat: Support for RS2 Downsampler (#465)
This implements the random selection from the [RS2
paper](https://openreview.net/pdf?id=JnRStoIuTe) minus the learning rate
scheduling adjustments.

Note that it is a bit suboptimal to use the downsampling infrastructure
here (#466). We might want to think about making the selector a bit more
dynamic, but for now, this will suffice to run experiments. #462 should
be merged before this is reviewed.