RHO-Loss Step 3: Train IL Model (#459)
This PR is the last PR to implement the RHO-LOSS downsampler on selector
side. After this PR, the whole process of training an IL model is
finished. Only a PR to use the IL model to train the main model, i.e.
`RemoteRHOLossDownsamplingStrategy` is missing.

The changes in this PR can be summarized as:
1. Extract the logic in `GRPCHandler` to talk with `trainer_server` as a
superclass `TrainerServerGRPCHandlerMixin`.
2. Extract the Pydantic Model `TrainingConfig` and place it in a
separate file, so that `RHOLossDownsamplingConfig` can reference it
(avoid circular import).
3. Use the capabilities of `TrainerServerGRPCHandlerMixin` in
`RHOLossDownsamplingStrategy` to train IL model.