Add description and scripts for downloading and generating the CGLM dataset (#474)
This adds a data generation script and some documentation on the CGLM
dataset. This can be further improved but for now gets the job done.

I also wanted to persist the scripts for scraping the metadata
somewhere, but they are not cleaned at all. I am okay with this, since I
do not want to put a lot of energy into that right now and we provide
the final scraped data anyways, so it's mostly for documentation
purposes (unless you want to scrape wikimedia for 2 weeks ;D).