Remove availability logic from trainer server (#486)
The current availability logic in the trainer server is a bit flawed. If
a training is running, the TS is considered unavailable. However, the
task of figuring out resources is currently left to the client anyways:
We specify the GPU and if we want to, we can start several pipelines as
long as another pipeline is currently not training (but instead, e.g.
evaluating).

This lead to some experiments not working out because the trainer server
not being available, although I wanted to train on a different GPU.
Let's remove this old logic and keep the full responsibility at client
for now. In the future, the TS should be responsible for managing its
resources, but as long as the client specifies the GPU, this logic does
not make sense.