feat: Warmup for data selection (#483)
One concern I have with evaluating data selection is that for `n`
triggers we should train on all the available data, and only then start
using data selection, to make the model adjust somewhat towards the new
data distribution first. This adds a warmup period to the downsampling
strategy, which disables presampling and downsampling during that
warmup.