More exact logging in storage and returning errors (#519)
I faced issues where for whatever reason in training the storage excited
because we tried to access a sample > 0 in a jpg. I have not encountered
the issue again, it might still be prevalent. This adds some more
detailed logging on what is happening for further investigation, and
also returns actual gRPC errors. I think the call propagation up the
chain might also not be working as expected currently, but I don't have
a log to test this with. Anyways, if you run into problems, having these
logs helps, so let's merge this into main.

Step 4/n of merging SIGMOD changes.