Fix recovery in datasets (#527)
Before, we had recovery logic based on reply indices. However, while
working on storage, I realized those responses come non
deterministically from multiple threads. Hence, we cannot rely on the
ordering. We need to keep track of the sample IDs we already yielded. I
changed the logic to just keep a list which is cheap to append to, and
only convert to a set / hash table as soon as we failed once and we
actually need to do many `in` checks.