fix: Adjust evaluation business logic (#535)
# Motivation

- We want to use the `real_last_timestamp` (start of next training
interval - 1 --> marking end of current training interval) only for
plotting the boxes in the heatmap plot. For decisions w.r.t. currently
active models this doesn't for. E.g. if the next year in a dataset has
no data at the year start, our training interval would extend into this
next year and therefore it's model won't be considered for the current
evaluation interval.
- timetriggers should allow starting a statically defined start point
(and not with the first sample), otherwise, the whole schedule is off a
couple of days.

# Note

Independently of the bug we fix with the `start_timestamp` in
`timetrigger`, this setting allows us to effectively do `pre-training`
with the first trigger (e.g. in the continous arxiv dataset we have
sparse data from 1988 - ~2005). We could simply start the schedule in
year 2005, then the first trigger trains on all previous data.