Fix batch number (#533)
Previously, we didn't record the number of passed batches correctly:
We use a `batch_number` which is generated purely from enumeration in
`dataloader`. Therefore this number is irrelevant to the number of
epochs (only shows how many batches there are in one epoch).
A similar issue exists on the iteration on `StB` when we calculate
scores class by class. The number of batches passed in the previous
class is not correctly accumulated on the current class.

This PR fixes it.