Parameterize `drop_last` (#550)
The reason for this PR is that, for il model training, I don't want to
drop the last batch, as the holdout set is already super small (and on
il model training no sampling is used, so it does not harm not to drop
last batch). But for main model training, usually we default `drop_last`
to True.