Batch evaluation intervals into a single request and a single evaluation process (#554)
Per the title, this PR enables batching evaluations on many intervals
into a single evaluation request. This resolves
https://github.com/eth-easl/modyn/issues/536. The integration test is
adjusted to cover this new functionality.

## How to review

The best way to review is to first take a look at
[modyn/protos/evaluator.proto](https://github.com/eth-easl/modyn/pull/554/files#diff-5017e582f2ffeabee6d3cfa42b2ca86f3f7035419df31cc3cff68be568279f04)
to see on the API level what is changed.

## Miscellaneous
After this PR, batching is only enabled on server side, on client side
still one interval is passed at a time, as I am confused on the current
way to generate `EvalRequest`.

After this PR, @robinholzi could you make a PR to collect the intervals
associated with the same `id_model` and `dataset_id`, and pack them in
one evaluation round? I think just looking at the function
`_single_batched_evaluation` in this file
[modyn/supervisor/internal/pipeline_executor/evaluation_executor.py](https://github.com/eth-easl/modyn/pull/554/files#diff-f111ee1d407da6a1b5d46b118dd5221da51801ead85ea2c9271b008299c6a11f)
should be enough to understand the change / how to make that PR. It
should be a very easy and straightforward PR. I am just confused with
the data models there. Thank you so much!

---------

Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>