Speed up `get_tensors_subset` from O(n^2) to O(n) (#623)
This is a change to replace the changes in my thesis branch to return
0-based indices in the `select_points()` method in downsamplers.

I realize that returning 0-based indices is not generalizable to all
downsamplers because there are downsamplers which work class by class
even in the batch-then-sample mode, e.g. craig. Therefore there is no
built-in concept of 0-based indices for such downsamplers (because they
pick points of the same class and select from them).

Therefore I decided to tweak the implementation of
`get_tensors_subset()` which is simpler.

A time complexity of O(n) is acceptable because in any selection
strategy we have to look at every sample, so an extra O(n) part can be
well mixed in the time complexity of a selection strategy.